requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
